pyproject.toml
[project]
name = "knowledge-graph-mcp"
version = "0.1.0"
description = "Persistent memory MCP server — Neo4j graph DB + vector embeddings"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.3.0",
    "neo4j>=5.26.0",
    "sentence-transformers>=3.0.1",
    "torch>=2.3.1",
    "python-dotenv>=1.0.1",
]

[project.scripts]
# Entry point — Claude Desktop calls this via:
#   uv run --directory /path/to/project knowledge-graph-mcp
knowledge-graph-mcp = "knowledge_graph_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge_graph_mcp"]

# ── uv configuration ──────────────────────────────────────────────────────────
# Use PyTorch CPU-only index to avoid downloading the full GPU build (~2GB).
# The CPU build (~700MB) is sufficient for embedding inference on this project.
[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url  = "https://download.pytorch.org/whl/cpu"
explicit = true
